testproj/requirements-github.txt
-e ..
drf-yasg==1.20.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
factory-boy==3.2.1